Conflicts:
	CommonWalkingControlModules/src/us/ihmc/commonWalkingControlModules/momentumBasedController/optimization/InverseDynamicsOptimizationControlModule.java
	CommonWalkingControlModules/src/us/ihmc/commonWalkingControlModules/momentumBasedController/optimization/MomentumOptimizationSettings.java
	IHMCRoboticsToolkit/src/us/ihmc/robotics/screwTheory/CompositeRigidBodyMassMatrixCalculator.java
	IHMCRoboticsToolkit/src/us/ihmc/robotics/screwTheory/InverseDynamicsJoint.java
	IHMCRoboticsToolkit/src/us/ihmc/robotics/screwTheory/ScrewTools.java
	IHMCRoboticsToolkit/src/us/ihmc/robotics/screwTheory/SpatialForceVector.java

+++ IHMCAvatarInterfaces/test/us/ihmc/avatar/testTools/DRCSimulationTestHelper.java
@@ -22,12 +22,14 @@ import us.ihmc.commonWalkingControlModules.configurations.WalkingControllerParam
 import us.ihmc.commonWalkingControlModules.desiredHeadingAndVelocity.HeadingAndVelocityEvaluationScriptParameters;
 import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.HighLevelBehaviorFactory;
 import us.ihmc.commonWalkingControlModules.highLevelHumanoidControl.factories.MomentumBasedControllerFactory;
+import us.ihmc.commons.PrintTools;
 import us.ihmc.communication.controllerAPI.command.Command;
 import us.ihmc.communication.net.LocalObjectCommunicator;
 import us.ihmc.communication.net.PacketConsumer;
 import us.ihmc.communication.packetCommunicator.PacketCommunicator;
 import us.ihmc.communication.packets.Packet;
 import us.ihmc.communication.util.NetworkPorts;
+import us.ihmc.euclid.geometry.BoundingBox3D;
 import us.ihmc.euclid.tuple3D.Point3D;
 import us.ihmc.euclid.tuple3D.Vector3D;
 import us.ihmc.humanoidBehaviors.behaviors.scripts.engine.ScriptBasedControllerCommandGenerator;
@@ -38,24 +40,22 @@ import us.ihmc.robotModels.FullHumanoidRobotModel;
 import us.ihmc.robotics.controllers.ControllerFailureException;
 import us.ihmc.robotics.dataStructures.registry.YoVariableRegistry;
 import us.ihmc.robotics.dataStructures.variable.YoVariable;
-import us.ihmc.robotics.geometry.BoundingBox3d;
 import us.ihmc.robotics.random.RandomGeometry;
 import us.ihmc.robotics.referenceFrames.ReferenceFrame;
 import us.ihmc.robotics.robotController.RobotController;
 import us.ihmc.robotics.robotSide.RobotSide;
 import us.ihmc.robotics.robotSide.SideDependentList;
 import us.ihmc.robotics.screwTheory.InverseDynamicsCalculatorListener;
+import us.ihmc.simulationConstructionSetTools.bambooTools.BambooTools;
+import us.ihmc.simulationConstructionSetTools.simulationTesting.NothingChangedVerifier;
+import us.ihmc.simulationConstructionSetTools.util.environments.CommonAvatarEnvironmentInterface;
+import us.ihmc.simulationConstructionSetTools.util.environments.DefaultCommonAvatarEnvironment;
 import us.ihmc.simulationconstructionset.HumanoidFloatingRootJointRobot;
 import us.ihmc.simulationconstructionset.SimulationConstructionSet;
-import us.ihmc.simulationconstructionset.bambooTools.BambooTools;
-import us.ihmc.simulationconstructionset.bambooTools.SimulationTestingParameters;
 import us.ihmc.simulationconstructionset.simulatedSensors.WrenchCalculatorInterface;
-import us.ihmc.simulationconstructionset.util.environments.CommonAvatarEnvironmentInterface;
-import us.ihmc.simulationconstructionset.util.environments.DefaultCommonAvatarEnvironment;
 import us.ihmc.simulationconstructionset.util.simulationRunner.BlockingSimulationRunner;
 import us.ihmc.simulationconstructionset.util.simulationRunner.BlockingSimulationRunner.SimulationExceededMaximumTimeException;
-import us.ihmc.simulationconstructionset.util.simulationTesting.NothingChangedVerifier;
-import us.ihmc.tools.io.printing.PrintTools;
+import us.ihmc.simulationconstructionset.util.simulationTesting.SimulationTestingParameters;
 import us.ihmc.tools.thread.ThreadTools;
 
 public class DRCSimulationTestHelper
@@ -81,46 +81,62 @@ public class DRCSimulationTestHelper
    private Exception caughtException;
 
    public DRCSimulationTestHelper(String name, DRCObstacleCourseStartingLocation selectedLocation,
-         SimulationTestingParameters simulationconstructionsetparameters, DRCRobotModel robotModel)
+                                  SimulationTestingParameters simulationconstructionsetparameters, DRCRobotModel robotModel)
    {
       this(new DefaultCommonAvatarEnvironment(), name, selectedLocation, simulationconstructionsetparameters, robotModel, true);
    }
 
-   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name,
-                                  DRCStartingLocation selectedLocation, SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel)
+   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel)
    {
       this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, true);
    }
 
-   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name,
-         DRCStartingLocation selectedLocation, SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel, boolean automaticallySpawnSimulation)
+   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel, boolean automaticallySpawnSimulation)
    {
-      this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, null, null, null, false, false, false, automaticallySpawnSimulation, null);
+      this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, null, null, null, false, false, false,
+           automaticallySpawnSimulation, null);
    }
 
-   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name,
-         DRCStartingLocation selectedLocation, SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel,
-         boolean addFootstepMessageGenerator, boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep, HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
+   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel, boolean addFootstepMessageGenerator,
+                                  boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep,
+                                  HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
    {
       this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, null, null, null, addFootstepMessageGenerator,
-            useHeadingAndVelocityScript, cheatWithGroundHeightAtForFootstep, true, walkingScriptParameters);
+           useHeadingAndVelocityScript, cheatWithGroundHeightAtForFootstep, true, walkingScriptParameters);
    }
 
    public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
-         SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel, DRCNetworkModuleParameters drcNetworkModuleParameters)
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel,
+                                  DRCNetworkModuleParameters drcNetworkModuleParameters)
    {
       this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, drcNetworkModuleParameters, null, null, false,
-            false, false, true, null);
+           false, false, true, null);
    }
 
    public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
-         SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel, DRCNetworkModuleParameters drcNetworkModuleParameters,
-         HighLevelBehaviorFactory highLevelBehaviorFactoryToAdd, DRCRobotInitialSetup<HumanoidFloatingRootJointRobot> initialSetup,
-         boolean addFootstepMessageGenerator, boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep,
-         boolean automaticallySpawnSimulation, HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel,
+                                  DRCNetworkModuleParameters drcNetworkModuleParameters, HighLevelBehaviorFactory highLevelBehaviorFactoryToAdd,
+                                  DRCRobotInitialSetup<HumanoidFloatingRootJointRobot> initialSetup, boolean addFootstepMessageGenerator,
+                                  boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep, boolean automaticallySpawnSimulation,
+                                  HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters)
+   {
+      this(commonAvatarEnvironmentInterface, name, selectedLocation, simulationTestingParameters, robotModel, drcNetworkModuleParameters,
+           highLevelBehaviorFactoryToAdd, initialSetup, addFootstepMessageGenerator, useHeadingAndVelocityScript, cheatWithGroundHeightAtForFootstep,
+           automaticallySpawnSimulation, walkingScriptParameters, true);
+   }
+   
+   public DRCSimulationTestHelper(CommonAvatarEnvironmentInterface commonAvatarEnvironmentInterface, String name, DRCStartingLocation selectedLocation,
+                                  SimulationTestingParameters simulationTestingParameters, DRCRobotModel robotModel,
+                                  DRCNetworkModuleParameters drcNetworkModuleParameters, HighLevelBehaviorFactory highLevelBehaviorFactoryToAdd,
+                                  DRCRobotInitialSetup<HumanoidFloatingRootJointRobot> initialSetup, boolean addFootstepMessageGenerator,
+                                  boolean useHeadingAndVelocityScript, boolean cheatWithGroundHeightAtForFootstep, boolean automaticallySpawnSimulation,
+                                  HeadingAndVelocityEvaluationScriptParameters walkingScriptParameters, boolean useBlockingSimulationRunner)
    {
       this.controllerCommunicator = PacketCommunicator.createIntraprocessPacketCommunicator(NetworkPorts.CONTROLLER_PORT,
-            new IHMCCommunicationKryoNetClassList());
+                                                                                            new IHMCCommunicationKryoNetClassList());
       this.testEnvironment = commonAvatarEnvironmentInterface;
 
       this.walkingControlParameters = robotModel.getWalkingControllerParameters();
@@ -174,9 +190,11 @@ public class DRCSimulationTestHelper
       scs = simulationStarter.getSimulationConstructionSet();
       sdfRobot = simulationStarter.getSDFRobot();
       avatarSimulation = simulationStarter.getAvatarSimulation();
-      blockingSimulationRunner = new BlockingSimulationRunner(scs, 60.0 * 10.0);
-      simulationStarter.attachControllerFailureListener(blockingSimulationRunner.createControllerFailureListener());
-
+      if (useBlockingSimulationRunner)
+      {
+         blockingSimulationRunner = new BlockingSimulationRunner(scs, 60.0 * 10.0);
+         simulationStarter.attachControllerFailureListener(blockingSimulationRunner.createControllerFailureListener());
+      }
 
       if (simulationTestingParameters.getCheckNothingChangedInSimulation())
       {
@@ -264,7 +282,7 @@ public class DRCSimulationTestHelper
    {
       return scriptedHandstepGenerator;
    }
-   
+
    public void checkNothingChanged()
    {
       if (simulationTestingParameters.getCheckNothingChangedInSimulation())
@@ -288,12 +306,18 @@ public class DRCSimulationTestHelper
 
    public void simulateAndBlock(double simulationTime) throws SimulationExceededMaximumTimeException, ControllerFailureException
    {
-      blockingSimulationRunner.simulateAndBlock(simulationTime);
+      if (blockingSimulationRunner != null)
+      {
+         blockingSimulationRunner.simulateAndBlock(simulationTime);
+      }
    }
 
    public void destroySimulation()
    {
-      blockingSimulationRunner.destroySimulation();
+      if (blockingSimulationRunner != null)
+      {
+         blockingSimulationRunner.destroySimulation();
+      }
       blockingSimulationRunner = null;
       if (avatarSimulation != null)
       {
@@ -336,7 +360,7 @@ public class DRCSimulationTestHelper
    {
       scs.getRootRegistry().addChild(childRegistry);
    }
-   
+
    public YoVariableRegistry getYovariableRegistry()
    {
       return scs.getRootRegistry();
@@ -379,16 +403,16 @@ public class DRCSimulationTestHelper
       scs.selectCamera("testCamera");
    }
 
-   public void assertRobotsRootJointIsInBoundingBox(BoundingBox3d boundingBox)
+   public void assertRobotsRootJointIsInBoundingBox(BoundingBox3D boundingBox)
    {
       assertRobotsRootJointIsInBoundingBox(boundingBox, getRobot());
    }
 
-   public static void assertRobotsRootJointIsInBoundingBox(BoundingBox3d boundingBox, HumanoidFloatingRootJointRobot robot)
+   public static void assertRobotsRootJointIsInBoundingBox(BoundingBox3D boundingBox, HumanoidFloatingRootJointRobot robot)
    {
       Point3D position = new Point3D();
       robot.getRootJoint().getPosition(position);
-      boolean inside = boundingBox.isInside(position);
+      boolean inside = boundingBox.isInsideInclusive(position);
       if (!inside)
       {
          fail("Joint was at " + position + ". Expecting it to be inside boundingBox " + boundingBox);
@@ -446,12 +470,13 @@ public class DRCSimulationTestHelper
 
       ArrayList<RobotController> footForceSensorSignalCorruptors = new ArrayList<RobotController>();
 
-      for(RobotSide robotSide : RobotSide.values)
+      for (RobotSide robotSide : RobotSide.values)
       {
-         for(int i = 0; i<footForceSensors.get(robotSide).size(); i++)
+         for (int i = 0; i < footForceSensors.get(robotSide).size(); i++)
          {
             ForceSensorHysteresisCreator forceSensorSignalCorruptor = new ForceSensorHysteresisCreator(sdfRobot.computeCenterOfMass(new Point3D()),
-                  footForceSensors.get(robotSide).get(i).getName(), footForceSensors.get(robotSide).get(i));
+                                                                                                       footForceSensors.get(robotSide).get(i).getName(),
+                                                                                                       footForceSensors.get(robotSide).get(i));
 
             footForceSensorSignalCorruptors.add(forceSensorSignalCorruptor);
          }
@@ -467,18 +492,16 @@ public class DRCSimulationTestHelper
 
       SideDependentList<String> jointNamesBeforeFeet = sdfRobot.getJointNamesBeforeFeet();
 
-      for(RobotSide robotSide : RobotSide.values)
+      for (RobotSide robotSide : RobotSide.values)
       {
-         footForceSensors.put(robotSide,new ArrayList<WrenchCalculatorInterface>());
-         for(int i = 0; i<forceSensors.size(); i++)
+         footForceSensors.put(robotSide, new ArrayList<WrenchCalculatorInterface>());
+         for (int i = 0; i < forceSensors.size(); i++)
          {
-            if(forceSensors.get(i).getJoint().getName().equals(jointNamesBeforeFeet.get(robotSide)))
+            if (forceSensors.get(i).getJoint().getName().equals(jointNamesBeforeFeet.get(robotSide)))
             {
                footForceSensors.get(robotSide).add(forceSensors.get(i));
             }
          }
       }
    }
-
-
 }
